chatbot-server/requirements.txt
chromadb
langchain
pip
openai
openapi-schema-pydantic
flask
# pydantic
websockets
uvicorn
Jinja2
unstructured
tabulate
Flask-Cors
pdf2image
tiktoken
gunicorn
six